Added the remixed changes
I added a copy of the entire project if you want to just add from unity hub (just rename the folder to what you want it to be), you could just copy the folder of only the assets and add inside of a existing project, but you would need to have the install all the packages, setup the project settings and would run in various errors, mainly with the URP.

I also made the assets folder available as a zip and there's even a unity package file you can directly import from a open unity project, that are all in my google drive.

If you are putting in a existing project I also recommend checking the packages required for the project (I will send them on discord but you can also see them in the full project folder). If your scene is all pink go in the project settings > URP and select the universal render pipeline asset, if you are having any other problems try reimport all the assets.